feat: added support for collecting handler result
Previous to this change it was only possible to collect the _last_ result returned from a handler (this is the default for `triggerHandler` which is what powers crank). This quickly becomes a problem when we want to deal with multiple async handlers as only the last promise would be accounted for.

To solve this we've introduced `collect` that you can use to wrap handlers like so:

```javascript
$('<div>")
  .on("test", collect(function() {
    return "first";
  })
  .on("test", collect(function() {
    return "last"
  })
  .triggerHandler("test"); // returns ["first", "last"]
```

BREAKING CHANGE: `jquery.crank` now accounts for usage of `collect` for async handler and unwraps arrays before running them through `jQuery.when`. If your handler returns an arrat you need to wrapp that in an outer array for it to work.